requirements.txt
chainlit==1.2.0
llama_index==0.11.13
pandas==2.2.3
pydantic==2.9.2
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
termcolor==2.4.0
yfinance==0.2.43
llama-index-llms-ollama
llama-index-tools-duckduckgo
llama-index-tools-bravesearch
llama-index-llms-groq
llama-index-llms-tavilysearch
llama-index-llms-huggingface
llama-index-llms-huggingface-api